pyproject.toml
[project]
name = "bumpit"
version = "0.1.0"
description = "Create CPFP (Child Pays For Parent) transactions for opening lightning channels in CLN"
readme = "README.md"
requires-python = ">=3.9.2"

dependencies = [
    "pyln-client>=24.11.1",
    "python-bitcointx==1.1.5",
    "python-bitcoinrpc==1.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4,<9",
    "pytest-xdist>=3.7,<4",
    "pytest-timeout>=2.4,<3",
    "pyln-testing>=24.11.1",
    "pyln-client>=24.11.1",
    "pyln-proto>=24.11.1",
]
